(refactor) OpusEncoder's _permitedFrameSize renamed to _permitedFrameSizeInMillisec, DEFAULT_AUDIO_FRAME_SIZE now a ushort, added documentation for all methods using the new SampleRate/SampleBits/SampleChannels/FrameRates arguments.

+++ MumbleSharp/Audio/AudioDecodingBuffer.cs
@@ -12,13 +12,21 @@ namespace MumbleSharp.Audio
         : IWaveProvider
     {
         private readonly int _sampleRate;
-        private readonly float _frameSize;
+        private readonly ushort _frameSize;
         public WaveFormat WaveFormat { get; private set; }
         private int _decodedOffset;
         private int _decodedCount;
         private readonly byte[] _decodedBuffer;
 
-        public AudioDecodingBuffer(int sampleRate = Constants.DEFAULT_AUDIO_SAMPLE_RATE, byte sampleBits = Constants.DEFAULT_AUDIO_SAMPLE_BITS, byte sampleChannels = Constants.DEFAULT_AUDIO_SAMPLE_CHANNELS, float frameSize = Constants.DEFAULT_AUDIO_FRAME_SIZE)
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="AudioDecodingBuffer"/> class which Buffers up encoded audio packets and provides a constant stream of sound (silence if there is no more audio to decode).
+        /// </summary>
+        /// <param name="sampleRate">The sample rate in Hertz (samples per second).</param>
+        /// <param name="sampleBits">The sample bit depth.</param>
+        /// <param name="sampleChannels">The sample channels (1 for mono, 2 for stereo).</param>
+        /// <param name="frameSize">Size of the frame in samples.</param>
+        public AudioDecodingBuffer(int sampleRate = Constants.DEFAULT_AUDIO_SAMPLE_RATE, byte sampleBits = Constants.DEFAULT_AUDIO_SAMPLE_BITS, byte sampleChannels = Constants.DEFAULT_AUDIO_SAMPLE_CHANNELS, ushort frameSize = Constants.DEFAULT_AUDIO_FRAME_SIZE)
         {
             WaveFormat = new WaveFormat(sampleRate, sampleBits, sampleChannels);
             _sampleRate = sampleRate;
@@ -142,7 +150,7 @@ namespace MumbleSharp.Audio
             //    _codec.Decode(null);
 
             var d = _codec.Decode(packet.Value.Data);
-            _nextSequenceToDecode = packet.Value.Sequence + d.Length / (int)(_sampleRate / _frameSize);
+            _nextSequenceToDecode = packet.Value.Sequence + d.Length / (_sampleRate / _frameSize);
 
             Array.Copy(d, 0, _decodedBuffer, _decodedOffset, d.Length);
             _decodedCount += d.Length;
