+++ MumbleSharp/Constants.cs
@@ -6,6 +6,6 @@ namespace MumbleSharp
         public const int DEFAULT_AUDIO_SAMPLE_RATE = 48000;
         public const byte DEFAULT_AUDIO_SAMPLE_BITS = 16;
         public const byte DEFAULT_AUDIO_SAMPLE_CHANNELS = 1;
-        public const byte DEFAULT_AUDIO_FRAME_SIZE = 20;
+        public const ushort DEFAULT_AUDIO_FRAME_SIZE = 20;
     }
 }
